+++ src/daemon/config/netdata-conf-backwards-compatibility.h
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_DAEMON_NETDATA_CONF_BACKWARDS_COMPATIBILITY_H
+#define NETDATA_DAEMON_NETDATA_CONF_BACKWARDS_COMPATIBILITY_H
+
+#include "config.h"
+
+void netdata_conf_backwards_compatibility(void);
+
+#endif //NETDATA_DAEMON_NETDATA_CONF_BACKWARDS_COMPATIBILITY_H
