+++ src/daemon/config/netdata-conf-db.h
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_DAEMON_NETDATA_CONF_DBENGINE_H
+#define NETDATA_DAEMON_NETDATA_CONF_DBENGINE_H
+
+#include "netdata-conf.h"
+
+void netdata_conf_section_db(void);
+
+#endif //NETDATA_DAEMON_NETDATA_CONF_DBENGINE_H
