+++ src/daemon/config/netdata-conf-directories.h
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_NETDATA_CONF_DIRECTORIES_H
+#define NETDATA_NETDATA_CONF_DIRECTORIES_H
+
+#include "netdata-conf.h"
+
+void netdata_conf_section_directories(void);
+
+#endif //NETDATA_NETDATA_CONF_DIRECTORIES_H
