+++ src/daemon/config/netdata-conf.h
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_DAEMON_NETDATA_CONF_H
+#define NETDATA_DAEMON_NETDATA_CONF_H
+
+#include "daemon/common.h"
+
+bool netdata_conf_load(char *filename, char overwrite_used, const char **user);
+
+#include "netdata-conf-backwards-compatibility.h"
+#include "netdata-conf-db.h"
+#include "netdata-conf-directories.h"
+#include "netdata-conf-global.h"
+#include "netdata-conf-logs.h"
+#include "netdata-conf-web.h"
+
+#endif //NETDATA_DAEMON_NETDATA_CONF_H
