+++ src/daemon/dyncfg/dyncfg-echo.c
