+++ src/daemon/dyncfg/dyncfg-inline.c
