+++ src/libnetdata/os/mem_available.h
@@ -5,6 +5,22 @@
 
 #include "libnetdata/libnetdata.h"
 
-uint64_t os_mem_available(void);
+typedef struct {
+    // ram_total_bytes is the total memory available in the system
+    // and it includes all the physical RAM the system may have.
+    // It does not include non-RAM memory (i.e. SWAP in not included).
+    // ram_total_bytes may be cached between calls to these functions
+    // (i.e. it may not be as up to date as ram_available_bytes).
+    uint64_t ram_total_bytes;
+
+    // ram_available_bytes is the total RAM memory available for
+    // applications, before the system and its applications run
+    // out-of-memory. This is always provided live by querying the system.
+    // It does not include non-RAM memory (i.e. SWAP in not included).
+    uint64_t ram_available_bytes;
+} OS_SYSTEM_MEMORY;
+
+// The function to get current system memory:
+OS_SYSTEM_MEMORY os_system_memory(bool query_total_ram);
 
 #endif //NETDATA_OS_MEM_AVAILABLE_H
